requirements.txt
typer
pydantic
requests
rich
python-dotenv
